moved source_urls logic to ResearchConductor
