+++ gpt_researcher/skills/researcher.py
@@ -38,8 +38,20 @@ class ResearchConductor:
             await stream_output("logs", "agent_generated", self.researcher.agent, self.researcher.websocket)
 
         # If specified, the researcher will use the given urls as the context for the research.
-        if self.researcher.source_urls:
-            self.researcher.context = await self.__get_context_by_urls(self.researcher.source_urls)
+        if self.source_urls:
+            self.context = await self.__get_context_by_urls(self.source_urls)
+            if len(self.context) == 0 and self.verbose:
+                # Could not find any relevant resources in source_urls to answer the query or sub-query. Will answer using model's inherent knowledge
+                await stream_output(
+                    "logs",
+                    "answering_from_memory",
+                    f"🧐 I was unable to find relevant context in the provided sources...",
+                    self.websocket,
+                )
+            # If add_additional_sources parameter is set, more resources can be gathered to create additional context using default web search
+            if self.add_additional_sources:
+                additional_research = await self.__get_context_by_search(self.query)
+                self.context += ' '.join(additional_research)
 
         elif self.researcher.report_source == ReportSource.Local.value:
             document_data = await DocumentLoader(self.researcher.cfg.doc_path).load()
